+++ onErrorHTTPResponse.go
@@ -7,17 +7,17 @@ import (
 	"github.com/xmidt-org/bascule/basculehttp"
 )
 
-type OnErrorHTTPResponseOption struct {
+type onErrorHTTPResponseConfig struct {
 	AuthType string
 }
 
-func onErrorHTTPResponse(config OnErrorHTTPResponseOption) (basculehttp.OnErrorHTTPResponse, error) {
+func onErrorHTTPResponse(config onErrorHTTPResponseConfig) (basculehttp.OnErrorHTTPResponse, error) {
 	if config.AuthType != "Bearer" && config.AuthType != "Basic" {
 		return nil, fmt.Errorf("invalid auth type '%s': expected Bearer or Basic", config.AuthType)
 	}
 	return func(w http.ResponseWriter, reason basculehttp.ErrorResponseReason) {
 		switch reason {
-		case basculehttp.ChecksNotFound, basculehttp.ChecksFailed:
+		case basculehttp.ChecksNotFound, basculehttp.ChecksFailed, basculehttp.ParseFailed:
 			w.WriteHeader(http.StatusForbidden)
 		default:
 			w.Header().Set(basculehttp.AuthTypeHeaderKey, config.AuthType)